restructure linter files
